Mixin update: - extracted Widgets builders into widgets/ - separated mixin into different files
